pyproject.toml
[tool.poetry]
name = "medicar"
version = "0.1.1"
description = ""
authors = ["raiizoor <mateus.informatica44@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
djangorestframework = "^3.12.4"
Django = "^3.2.7"
django-filter = "^2.4.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
